log_userwarning verplaatst naar _logging.py
